moving CallToActionBlockContainer into component folder, and organizing component
